Moved AssociationGroupTestingMixin for general usage
